chore: update Magic.NET-Q8-AnyCPU to 14.9.1

+++ Shoko.Server/Models/Image_Base.cs
@@ -278,8 +278,8 @@ public class Image_Base : IImageMetadata
                 return;
             }
 
-            _width = info.Width;
-            _height = info.Height;
+            _width = (int)info.Width;
+            _height = (int)info.Height;
         }
         catch
         {
